requirements.txt
bs4
random
wordcloud
Pillow
requests
jieba
numpy